+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Added
+* Added `Animation::skinning_transforms` and `Animation::model_space_transforms` to xc3_model.
+
+### Changed
+* Renamed `Skeleton::world_transforms` to `Skeleton::model_space_transforms` for xc3_model.
+
 ## 0.7.0 - 2024-03-15
 ### Added
 * Added support for anisotropic filtering to xc3_wgpu.